requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-stubs==4.2.6
django-stubs-ext==4.2.5
djangorestframework==3.14.0
idna==3.4
influxdb-client==1.38.0
mysqlclient==2.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
reactivex==4.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.6.0
